deleted pages components
