--- src/Components/Pages/HomePage/index.tsx
@@ -1,11 +1,0 @@
-import React, { FC } from 'react';
-import { ContactList, Header } from '../../index';
-
-const HomePage: FC = () => (
-  <>
-    <Header />
-    <ContactList />
-  </>
-);
-
-export default HomePage;
