+++ src/Components/index.ts
@@ -1,9 +1,7 @@
 export { default as Header } from './Header';
 export { default as ContactList } from './ContactList';
-export { default as ErrorPage } from './Pages/ErrorPage';
-export { default as ContactPage } from './Pages/ContactPage';
-export { default as LeadersPage } from './Pages/LeadersPage';
+export { default as ErrorPage } from './DisplayError';
+export { default as FullContact } from './FullContact';
 export { default as PreLoader } from './Preloader';
-export { default as HomePage } from './Pages/HomePage';
 export { default as LeadersList } from './LeadersList';
 export { default as RatingList } from './RaitingList';
